split ViewManager sample into 2 to separate out PanelSet port
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)
